use signed int for thread id

+++ src/dll_inject.cpp
@@ -33,7 +33,7 @@ void dll_inject::suspend_inject_resume(
     const std::chrono::milliseconds delay_pre_resume) {
   process::Process P(ex_handle);
   P.suspend_threads(-1);
-  unsigned long tid = 0;
+  int tid = 0;
   P.for_each_thread([&tid](process::Thread* T, void* ctx) {
     (void)ctx;
     if (tid == 0) {
